% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_corpus.R
\name{clean_corpus}
\alias{clean_corpus}
\title{clean_corpus Function}
\usage{
clean_corpus(corpus)
}
\arguments{
\item{corpus}{input corpus}
}
\description{
This function allows you to clean corpus.
}
\examples{
clean-corpus(corpus)
}
\keyword{corpus}
