% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{df_to_matrix Function}
\usage{
df_to_matrix(data)
}
\arguments{
\item{data}{Input dataframe}
}
\description{
This function allows you to convert a dataframe to a matrix.
}
\keyword{convert}
