% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_id.R
\name{get_product_id}
\alias{get_product_id}
\title{get_product_id Function}
\usage{
get_product_id(user_input_text, full_data, N)
}
\arguments{
\item{user_input_text}{User's input text}

\item{full_data}{Product database}

\item{N}{length limitation of the results}
}
\description{
This function allows you to get a list of product_id according to user's input text.
}
\examples{
get_product_id("I need a comfortable dress", read.csv("sentiment_df_v1"), 3 )
}
\keyword{userinput}
